Code refactored
